ゲーム終了時点数など正しく表示、保存完了

+++ Assets/Script/DatabaseManager.cs
@@ -54,7 +54,7 @@ public class DatabaseManager : MonoBehaviour
     }
 
     // ÉfÅ[É^ÇÃí«â¡
-    public void AddResult(int point, int typingCount,float accuracy, int speed)
+    public void AddResult(int point, int typingCount,float accuracy, float speed)
     {
         var typingResult = new TypingResult
         {
